requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alpaca-py==0.43.2
alpaca-trade-api==3.2.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
deprecation==2.1.0
exceptiongroup==1.3.1
flake8==7.3.0
frozenlist==1.8.0
idna==3.11
iniconfig==2.1.0
joblib==1.5.3
mccabe==0.7.0
msgpack==1.0.3
multidict==6.7.1
numpy==2.0.2
packaging==26.1
pandas==2.3.3
pluggy==1.6.0
propcache==0.4.1
pyarrow==21.0.0
pycodestyle==2.14.0
pydantic==2.13.3
pydantic_core==2.46.3
pyflakes==3.4.0
Pygments==2.20.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.1
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sseclient-py==1.9.0
threadpoolctl==3.6.0
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==1.26.20
websocket-client==1.9.0
websockets==10.4
yarl==1.22.0
